requirements.txt
Babel==2.5.1
blinker==1.4
certifi==2017.11.5
chardet==3.0.4
Click==7.0
cycler==0.10.0
defusedxml==0.5.0
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
gunicorn==19.6.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10.1 #updated 2019-05-30
MarkupSafe==1.0
#matplotlib==1.4.3
nose==1.3.7
numpy==1.18.4
pandas==1.0.3
pyparsing==2.2.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
requests==2.20.0
#requirements==0.1
scipy==1.4.1
six==1.11.0
SQLAlchemy==1.3.4 #updated 2019-05-30
urllib3==1.25.3 #updated 2019-05-30
Werkzeug==0.16.0 #updated 2020-01-10
htthttplib2==0.18.1 #updated 2020-08-28


#updated 2020-08-28
cachetools=4.1.1
google-api-core=1.22.1
google-api-python-client=1.11.0
google-auth=1.21.0
google-auth-httplib2=0.0.4
googleapis-common-protos=1.52.0
protobuf=3.13.0
pyasn1=0.4.8
pyasn1-modules=0.2.8
rsa=4.6
uritemplate=3.0.1
